Test suite

+++ app/models/category.rb
@@ -1,5 +1,10 @@
 class Category
 
+	# A Category is a label applied to a group of related Codes within a provided context.
+	# For example, a category may refer to a subset of the codes related to "age".
+	# Categories are machine-derived. As such, they are influenced by biases in external training data.
+	# Careful human discernment of categories is required to identify and address these biases.
+	
 	include ActiveGraph::Node
 	
 	property :name
@@ -10,7 +15,7 @@ class Category
 
 	has_many :out, :codes, rel_class: :CategorizedAs, dependent: :delete_orphans
 
-	PROMPT_INITIALIZE = %{ 
+	PROMPT = %{ 
 		You are a social researcher doing data analysis. Please generate a list of the 20 most relevant themes from the following list of codes. The themes should be all lowercase and contain no punctuation. Codes should be stripped of quotation marks. Return each code with an array of its categories in JSON format. Use this JSON as the format:
 		
 		{ 
@@ -25,34 +30,25 @@ class Category
 		The codes are as follows: 
 	}
 
-	def self.enqueue_category_extractor_job(context)
-		CategoryExtractorJob.perform_async(context)
-	end
-
-	def self.from_context(context)
+	def self.from(context)
 		codes = Code.where(context: context)
-		client = OpenAI::Client.new
-	
-		response = client.chat(
-			parameters: {
-				model: "gpt-4o",
-				response_format: { type: "json_object" },
-				messages: [{ role: "user", content: "#{PROMPT_INITIALIZE} #{codes.map(&:name).join(",")}" }],
-				temperature: 0.7,
-			}
-		)	
-
-		data = JSON.parse(response.dig("choices", 0, "message", "content"))['themes']
-
+		response = Clients::OpenAi.request("#{PROMPT} #{codes.map(&:name).join(',')}")
+		return unless response['themes']
+		
 		Category.where(context: context).destroy_all
 
-		data.each do |record|
+		response['themes'].each do |record|
 			category = Category.find_or_create_by(name: record['theme'], context: context)
-			record['codes'].each do |v|
-				codes.select{ |code| record['codes'].include? code.name }.each{ |code| CategorizedAs.create(from_node: category, to_node: code )}
+			codes.each do |code|
+				next unless record['codes'].include?(code.name)
+				CategorizedAs.create(from_node: category, to_node: code)
 			end
 		end
 
 	end
 
+	def self.histogram(context)
+		where(context: context).inject({}) { |acc, category| acc[category.name] = category.codes.count; acc }
+	end
+	
 end
